+++ ontimize-jee-report-common/src/main/java/com/ontimize/jee/report/common/util/TypeMappingsUtils.java
@@ -1,5 +1,6 @@
 package com.ontimize.jee.report.common.util;
 
+import java.math.BigDecimal;
 import java.sql.Types;
 import java.util.Collection;
 import java.util.Date;
@@ -16,13 +17,17 @@ public class TypeMappingsUtils {
             DOUBLE = "Double",
             DATE = "Date",
             INTEGER = "Integer",
+            BIGDECIMAL = "BigDecimal",
 
     STRING_PATH = "java.lang.String",
             BOOLEAN_PATH = "java.lang.Boolean",
             OBJECT_PATH = "java.lang.Object",
+            FLOAT_PATH = "java.lang.Float",
             DOUBLE_PATH = "java.lang.Double",
             DATE_PATH = "java.util.Date",
-            INTEGER_PATH = "java.lang.Integer";
+            TIMESTAMP_PATH = "ava.sql.Timestamp",
+            INTEGER_PATH = "java.lang.Integer",
+            BIGDECIMAL_PATH = "java.math.BigDecimal";
 
     @SuppressWarnings("rawtypes")
     public static Class getClass(int type) {
@@ -45,9 +50,11 @@ public class TypeMappingsUtils {
             case Types.DOUBLE:
             case Types.DECIMAL:
             case Types.REAL:
-            case Types.NUMERIC:
             case Types.FLOAT:
                 return Double.class;
+
+            case Types.NUMERIC:
+                return BigDecimal.class;
 
             case Types.DATE:
             case Types.TIME:
@@ -90,9 +97,11 @@ public class TypeMappingsUtils {
             case Types.DOUBLE:
             case Types.DECIMAL:
             case Types.REAL:
-            case Types.NUMERIC:
             case Types.FLOAT:
                 return DOUBLE_PATH;
+
+            case Types.NUMERIC:
+                return BIGDECIMAL_PATH;
 
             case Types.DATE:
             case Types.TIME:
@@ -114,6 +123,36 @@ public class TypeMappingsUtils {
         return getClassName(getSQLType(type));
     }
 
+    public static int getSQLTypeFromClassName(final String className) {
+        switch (className) {
+            case STRING_PATH:
+                return Types.VARCHAR;
+
+            case INTEGER_PATH:
+                return Types.INTEGER;
+
+            case BOOLEAN_PATH:
+                return Types.BOOLEAN;
+
+            case FLOAT_PATH:
+                return Types.FLOAT;
+
+            case DOUBLE_PATH:
+                return Types.DOUBLE;
+
+            case BIGDECIMAL_PATH:
+                return Types.NUMERIC;
+
+            case DATE_PATH:
+                return Types.DATE;
+
+            case TIMESTAMP_PATH:
+                return Types.TIMESTAMP;
+
+        }
+        return Types.OTHER;
+    }
+
     public static int getSQLType(String type) {
         int returned = Types.OTHER;
         if (type.equalsIgnoreCase(STRING)) {
@@ -124,7 +163,9 @@ public class TypeMappingsUtils {
             returned = Types.BINARY;
         } else if (type.equalsIgnoreCase(DOUBLE)) {
             returned = Types.DOUBLE;
-        } else if (type.equalsIgnoreCase(DATE)) {
+        } else if (type.equalsIgnoreCase(BIGDECIMAL)) {
+            returned = Types.NUMERIC;
+        }else if (type.equalsIgnoreCase(DATE)) {
             returned = Types.DATE;
         } else if (type.equalsIgnoreCase(INTEGER)) {
             returned = Types.INTEGER;