+++ CHANGELOG.md
@@ -7,6 +7,48 @@
 <!-- ### Security 🛡️-->
 
 ## [Unreleased]
+## [3.5.0] - 2025-04-28
+### Fixed 🐛
+* **Report store**: Fixed problems when using parameters of Jasper Report. It was added new `ReportStoreParamsDto` object into `fillReport` API method for allowing customization of parameters as types, formattings, etc.
+* **Report on demand**: Fixed problem when using BigDecimals when filling reports.
+### Breaking changes
+* The `/fillReport/{id}` method of `ReportStoreRestController` has changed from:
+```java
+public EntityResult fillReport(@PathVariable("id") String id,
+                                   @RequestBody(required = true) Map<String, Object> bodyParams)
+```
+to:
+```java
+public EntityResult fillReport(@PathVariable("id") String id,
+                                   @RequestBody(required = true) ReportStoreParamsDto bodyParams)
+```
+
+* The `buildReport` method of `IDynamicJasperService` interface has changed from:
+```java
+public DynamicReport buildReport(List<ColumnDto> columns, String title, List<String> groups, String entity,
+                                 String service, String path, Boolean vertical, List<FunctionTypeDto> functions, StyleParamsDto styles, String subtitle,
+                                 String language)
+        throws DynamicReportException;
+```
+to:
+```java
+public DynamicReport buildReport(ReportParamsDto reportParamsDto) throws DynamicReportException;
+```
+
+* The `generateReport` method of `ReportBase` abstract class has changed from:
+```java
+public InputStream generateReport(List<ColumnDto> columns, String title, List<String> groups, String entity,
+            String service, String path, Boolean vertical, List<FunctionTypeDto> functions, StyleParamsDto style,
+            String subtitle, List<OrderByDto> orderBy, String language, FilterParameter filters, Boolean advQuery)
+            throws DynamicReportException
+```
+to:
+```java
+public InputStream generateReport(final ReportParamsDto reportParamsDto)
+        throws DynamicReportException
+```
+
+## [3.4.0] - 2024-02-23
 ### Added ✔️
 * **JaCoCo**: Add JaCoCo coverage on Sonar
 ### Changed 🛠️
@@ -24,6 +66,8 @@
 ### Added ✔️
 * **Report on-demand**: On-demand reports can now be created with the help of Jasper Reports via HTTP requests. The parameters for creating these reports can be stored in preferences for later use.
 
-[unreleased]: https://github.com/ontimize/ontimize-jee-report/compare/3.3.0...HEAD
+[unreleased]: https://github.com/ontimize/ontimize-jee-report/compare/3.5.0...HEAD
+[3.5.0]: https://github.com/ontimize/ontimize-jee-report/compare/3.4.0...3.5.0
+[3.4.0]: https://github.com/ontimize/ontimize-jee-report/compare/3.3.0...3.4.0
 [3.3.0]: https://github.com/ontimize/ontimize-jee-report/compare/3.2.0...3.3.0
 [3.2.0]: https://github.com/ontimize/ontimize-jee-report/tree/3.2.0